requirements.txt
torch==1.13.1
numpy==1.24.0
opencv-python==4.6.0.66
torchvision==0.14.1
Pillow==9.3.0
tqdm==4.64.1
diffusers==0.11.1
Django==4.1.4